topic/config-cleanup: 	- mcp23017 cleanup
